backend/requirements.txt
# Core production dependencies
fastapi[standard]==0.116.1
uvicorn[standard]==0.31.0
python-multipart==0.0.18
python-dotenv==1.0.1
httpx==0.27.0
openai==1.99.4
pydantic==2.11.7
starlette==0.36.1

# Optional: Authentication and file handling (if needed later)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
